+++ test/WebJobs.Script.Tests/FunctionMetadataProviderTests.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.IO.Abstractions.TestingHelpers;
 using System.Linq;
+using Microsoft.Azure.WebJobs.Script.Description;
 using Microsoft.Azure.WebJobs.Script.Diagnostics;
 using Microsoft.Azure.WebJobs.Script.Workers.Rpc;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -41,6 +42,31 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
 
             Assert.Equal(18, metadataProvider.GetFunctionMetadata(false).Length);
             Assert.True(AreRequiredMetricsEmitted(_testMetricsLogger));
+        }
+
+        [Fact]
+        public void ReadFunctionMetadata_With_Retry_Succeeds()
+        {
+            string functionsPath = Path.Combine(Environment.CurrentDirectory, @"..\..\..\..\..\sample\noderetry");
+            _scriptApplicationHostOptions.ScriptPath = functionsPath;
+            var optionsMonitor = TestHelpers.CreateOptionsMonitor(_scriptApplicationHostOptions);
+            var metadataProvider = new FunctionMetadataProvider(optionsMonitor, new OptionsWrapper<LanguageWorkerOptions>(_languageWorkerOptions), NullLogger<FunctionMetadataProvider>.Instance, _testMetricsLogger);
+            var workerConfigs = TestHelpers.GetTestWorkerConfigs();
+            var functionMetadatas = metadataProvider.GetFunctionMetadata(false);
+
+            Assert.Equal(2, functionMetadatas.Length);
+
+            var functionMetadataWithRetry = functionMetadatas.Where(f => f.Name.Contains("HttpTrigger-RetryFunctionJson", StringComparison.OrdinalIgnoreCase));
+            Assert.Single(functionMetadataWithRetry);
+            var retry = functionMetadataWithRetry.FirstOrDefault().Retry;
+            Assert.NotNull(retry);
+            Assert.Equal(RetryStrategy.FixedDelay, retry.Strategy);
+            Assert.Equal(4, retry.MaxRetryCount);
+            Assert.Equal(TimeSpan.Parse("00:00:03"), retry.DelayInterval);
+
+            var functionMetadata = functionMetadatas.Where(f => !f.Name.Contains("HttpTrigger-RetryFunctionJson", StringComparison.OrdinalIgnoreCase));
+            Assert.Single(functionMetadataWithRetry);
+            Assert.Null(functionMetadata.FirstOrDefault().Retry);
         }
 
         private bool AreRequiredMetricsEmitted(TestMetricsLogger metricsLogger)